moved in-memory network to bootstrap package
